refactor(frontend): extract demo asset minting
