+++ packages/demo/frontend/src/components/earn/EarnWithFrontendWallet.tsx
@@ -1,11 +1,7 @@
-import { encodeFunctionData, type Address } from 'viem'
 import type {
   Wallet,
-  SupportedChainId,
-  Asset,
   ReactProviderTypes,
 } from '@eth-optimism/actions-sdk/react'
-import { mintableErc20Abi } from '@/abis/mintableErc20Abi'
 import { getBlockExplorerUrl } from '@/utils/blockExplorer'
 import Earn from './Earn'
 import {
@@ -16,8 +12,8 @@ import {
 import { useMemo } from 'react'
 import { createActions } from '@eth-optimism/actions-sdk/react'
 import { createActionsConfig } from '@/config/actions'
-import { actionsApi } from '@/api/actionsApi'
 import type { EarnOperations } from '@/hooks/useLendProvider'
+import { mintDemoAsset } from '@/utils/demoAssetMinting'
 
 export interface EarnWithFrontendWalletProps {
   wallet: Wallet | null
@@ -52,54 +48,7 @@ export function buildFrontendWalletOperations(
     getTokenBalances: async () => wallet.getBalance(),
     getMarkets: async () => actions.lend.getMarkets(),
     getPosition: async (marketId) => wallet.lend!.getPosition({ marketId }),
-    mintAsset: async (asset: Asset) => {
-      const walletAddress = wallet.address
-      const chainId = asset.address
-        ? Object.keys(asset.address).find(
-            (key) => asset.address[key as unknown as SupportedChainId],
-          )
-        : undefined
-
-      if (!chainId) throw new Error('No chain available for asset')
-
-      if (asset.metadata.symbol === 'ETH' && asset.type === 'native') {
-        await actionsApi.dripEthToWallet(walletAddress)
-        return
-      }
-
-      const amountInDecimals = BigInt(
-        Math.floor(parseFloat('100') * Math.pow(10, asset.metadata.decimals)),
-      )
-      const tokenAddress = asset.address[parseInt(chainId) as SupportedChainId]
-
-      if (!tokenAddress || tokenAddress === 'native') {
-        throw new Error(
-          `Asset ${asset.metadata.symbol} not available on chain ${chainId}`,
-        )
-      }
-
-      const result = await wallet.sendBatch(
-        [
-          {
-            to: tokenAddress as Address,
-            data: encodeFunctionData({
-              abi: mintableErc20Abi,
-              functionName: 'mint',
-              args: [walletAddress, amountInDecimals],
-            }),
-            value: 0n,
-          },
-        ],
-        parseInt(chainId) as SupportedChainId,
-      )
-
-      if ('blockExplorerUrl' in result && result.blockExplorerUrl) {
-        return { blockExplorerUrls: [result.blockExplorerUrl as string] }
-      }
-      if ('blockExplorerUrls' in result && result.blockExplorerUrls) {
-        return { blockExplorerUrls: result.blockExplorerUrls as string[] }
-      }
-    },
+    mintAsset: async (asset) => mintDemoAsset(wallet, asset),
     openPosition: async (params) => wallet.lend!.openPosition(params),
     closePosition: async (params) => wallet.lend!.closePosition(params),
     executeSwap: async (quote) => {
